backend/requirements.txt
pip
setuptools
wheel
fastapi==0.110.0
uvicorn==0.29.0
yfinance==0.2.37
pandas==1.5.3
numpy==1.24.4
requests==2.31.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.29
google-generativeai==0.4.0
google-auth==2.29.0
httpx==0.28.1
email-validator>=2.0.0
sendgrid>=6.12.5
finnhub-python>=2.4.27
nsepython>=0.0.12
PyJWT==2.8.0